+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## XX.XX.XX
 * The feedback widgets now have transparent backgrounds for a cleaner look.
 * Extended the notification button URL handler to allow custom handling of URLs when notification buttons are clicked in the background.
-* Added a config method to disable server config in the initialization "disableSDKBehaviorSettings()".
+* Added a config method to disable server config updates in the initialization "disableSDKBehaviorSettingsUpdates()".
 
 * Deprecated "presentFeedbackWidget(widgetInfo, context, closeButtonText, devCallback)", replaced with "presentFeedbackWidget(widgetInfo, context, devCallback)" in the feedbacks.